+++ plugin/src/main/java/de/infolektuell/gradle/jextract/GradleJextractPlugin.java
@@ -3,27 +3,39 @@ package de.infolektuell.gradle.jextract;
 import de.infolektuell.gradle.jextract.extensions.JextractExtension;
 import de.infolektuell.gradle.jextract.extensions.SourceSetExtension;
 import de.infolektuell.gradle.jextract.service.JextractStore;
-import de.infolektuell.gradle.jextract.tasks.JextractBaseTask;
+import de.infolektuell.gradle.jextract.tasks.*;
+
 import static de.infolektuell.gradle.jextract.tasks.JextractBaseTask.*;
-import de.infolektuell.gradle.jextract.tasks.JextractDumpIncludesTask;
-import de.infolektuell.gradle.jextract.tasks.JextractGenerateTask;
+
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
+import org.gradle.api.attributes.LibraryElements;
 import org.gradle.api.file.FileCollection;
+import org.gradle.api.plugins.JavaApplication;
 import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.JavaPluginExtension;
 import org.gradle.api.provider.Provider;
+import org.gradle.api.tasks.JavaExec;
 import org.gradle.api.tasks.TaskProvider;
+import org.gradle.api.tasks.testing.Test;
 import org.gradle.internal.jvm.Jvm;
+import org.gradle.jvm.toolchain.JavaCompiler;
 import org.gradle.jvm.toolchain.JavaLanguageVersion;
+import org.gradle.jvm.toolchain.JavaToolchainService;
 import org.jspecify.annotations.NonNull;
 
+import javax.inject.Inject;
+import java.io.File;
+import java.nio.file.FileSystems;
+import java.nio.file.Files;
+import java.nio.file.PathMatcher;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
 
 /// A Gradle plugin to add Jextract to the build
-public class GradleJextractPlugin implements Plugin<@NonNull Project> {
+@SuppressWarnings("UnstableApiUsage")
+public abstract class GradleJextractPlugin implements Plugin<@NonNull Project> {
     /// The plugin ID that must be used in build scripts to apply the plugin
     public static final String PLUGIN_NAME = "de.infolektuell.jextract";
 
@@ -65,11 +77,18 @@ public class GradleJextractPlugin implements Plugin<@NonNull Project> {
                 task.getInstallation().convention(jextractInstallation);
             });
 
+            final LibraryPathProvider libraryPathProvider = project.getObjects().newInstance(LibraryPathProvider.class);
+            project.getTasks().withType(Test.class, task -> {
+                task.getJvmArgumentProviders().add(libraryPathProvider);
+            });
+
             final Map<String, TaskProvider<@NonNull JextractGenerateTask>> jextractGenerateTasks = new HashMap<>();
             extension.getLibraries().configureEach(lib -> {
                 lib.getUseSystemLoadLibrary().convention(false);
                 lib.getOutput().convention(extension.getOutput().dir(lib.getName()));
                 lib.getGenerateSourceFiles().convention(extension.getGenerateSourceFiles());
+                lib.getIncludes().add(lib.getHeader().map(h -> project.getLayout().getProjectDirectory().dir(h.getAsFile().getParentFile().getAbsolutePath())));
+                libraryPathProvider.getFiles().from(lib.getLibraryPath());
 
                 final TaskProvider<@NonNull JextractGenerateTask> generateTaskProvider = project.getTasks().register(lib.getName() + "JextractGenerateBindings", JextractGenerateTask.class, task -> {
                     task.setDescription("Uses Jextract to generate Java bindings for the " + lib.getName() + " native library");
@@ -111,6 +130,74 @@ public class GradleJextractPlugin implements Plugin<@NonNull Project> {
                     s.setRuntimeClasspath(s.getRuntimeClasspath().plus(src));
                 });
             });
+            javaExtension.getSourceSets().named("main", s -> {
+                final var isModularProject = s.getJava().getSourceDirectories().filter(this::isModule).getElements().map(e -> !e.isEmpty());
+                final TaskProvider<@NonNull JmodCreateTask> createJmodTask = project.getTasks().register(s.getTaskName("create", "Jmod"), JmodCreateTask.class, task -> {
+                    task.setGroup("build");
+                    task.setDescription("Converts a jar into a jmod file");
+                    task.onlyIf(t -> isModularProject.get());
+                    task.getMetadata().convention(javaVersion.flatMap(v -> getJavaToolchainService().compilerFor(spec -> spec.getLanguageVersion().set(v)).map(JavaCompiler::getMetadata)));
+                    task.getClasspath().from(s.getOutput());
+                    task.getJmod().convention(project.getLayout().getBuildDirectory().file(String.format("libs/%s.jmod", project.getName())));
+                });
+
+                project.getConfigurations().named(s.getApiElementsConfigurationName(), config -> {
+                    config.getOutgoing().getVariants().register("jmod", jmod -> {
+                        jmod.getDescription().set("A jmod file containing classes, resources, libs, headers, and legal notices if available.");
+                        jmod.getAttributes().attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, project.getObjects().named(LibraryElements.class, "jmod"));
+                        jmod.artifact(createJmodTask, artifact -> artifact.setType("jmod"));
+                    });
+                });
+                project.getConfigurations().named(s.getRuntimeElementsConfigurationName(), config -> {
+                    config.getOutgoing().getVariants().register("jmod", jmod -> {
+                        jmod.getDescription().set("A jmod file containing classes, resources, libs, headers, and legal notices if available.");
+                        jmod.getAttributes().attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, project.getObjects().named(LibraryElements.class, "jmod"));
+                        jmod.artifact(createJmodTask, artifact -> artifact.setType("jmod"));
+                    });
+                });
+
+                final SourceSetExtension sourceSetExtension = s.getExtensions().getByType(SourceSetExtension.class);
+                sourceSetExtension.getLibraries().all(lib -> {
+                    createJmodTask.configure(task -> {
+                        task.getHeaderFiles().from(lib.getIncludes());
+                        task.getLibs().from(lib.getLibraryPath());
+                    });
+                });
+            });
         });
+
+        project.getPluginManager().withPlugin("application", applicationPlugin -> {
+            final JavaPluginExtension javaExtension = project.getExtensions().getByType(JavaPluginExtension.class);
+            final JavaApplication application = project.getExtensions().getByType(JavaApplication.class);
+            javaExtension.getSourceSets().named("main", s -> {
+                project.getTasks().named(s.getTaskName("create", "jmod"), JmodCreateTask.class, task -> {
+                    task.getMainClass().convention(application.getMainClass());
+                });
+            });
+
+            final LibraryPathProvider libraryPathProvider = project.getObjects().newInstance(LibraryPathProvider.class);
+            extension.getLibraries().all(lib -> {
+                libraryPathProvider.getFiles().from(lib.getLibraryPath());
+            });
+            project.getTasks().named("run", JavaExec.class, task -> {
+                task.getJvmArgumentProviders().add(libraryPathProvider);
+            });
+        });
+    }
+
+    ///  Injects the build service for Java toolchains
+    /// @return A property holding the injected build service
+    @Inject
+    protected abstract JavaToolchainService getJavaToolchainService();
+
+    private boolean isModule(File file) {
+        if (file.isFile() && file.getName().endsWith(".jmod")) return true;
+        if (!file.isDirectory()) return false;
+        PathMatcher matcher = FileSystems.getDefault().getPathMatcher("glob:**/module-info.{class,java}");
+        try (var stream = Files.walk(file.toPath(), 2)) {
+            return stream.anyMatch(matcher::matches);
+        } catch (Exception ignored) {
+            return false;
+        }
     }
 }